requirements.txt
discord.py
discord-ext-voice-recv
audioop-lts
PyNaCl
python-dotenv